+++ lib/emcp/resource.ex
@@ -5,7 +5,7 @@ defmodule EMCP.Resource do
   @callback name() :: String.t()
   @callback description() :: String.t()
   @callback mime_type() :: String.t()
-  @callback read() :: String.t()
+  @callback read(conn :: Plug.Conn.t() | nil) :: String.t()
 
   def to_map(module) do
     %{
@@ -16,7 +16,7 @@ defmodule EMCP.Resource do
     }
   end
 
-  def to_contents(module) do
-    [%{"uri" => module.uri(), "mimeType" => module.mime_type(), "text" => module.read()}]
+  def to_contents(module, conn) do
+    [%{"uri" => module.uri(), "mimeType" => module.mime_type(), "text" => module.read(conn)}]
   end
 end